WIP. Refactor download. Cleanup.

+++ vgrab.py
@@ -3,24 +3,56 @@ from lxml import html,etree
 import os,sys
 import urllib2
 import urllib
+import argparse
 
+#import codecs
+#codecs.register(lambda name: codecs.lookup('utf-8') if name == 'cp65001' else None)
 
 # FIXME:
 # ======
 #
-# - Implement argparse
 # - Download "vedlagte miniatyrbilder". See post 77
 # - Change download to fetch local cache first, then try server
-# - Do progress printing to stderr
-#
-
+# - Handle caching if default number of posts per page changes
 
 START_URL = 'http://avforum.no/forum/showthread.php?t=126778'
 
 USE_TMP=True
-SLURP='tmp/'
 
 POST_PER_PAGE=50
+
+# Images that are used from e.g. CSS and needs to be pulled manually
+EXTRA_IMAGES = (
+    'images/misc/quote-left.png',
+)
+
+def progress(n,max,e=''):
+
+    l=80
+    p = n*80/max
+    s1 = '=' * p
+    s2 = ' ' * (l-p)
+    print >>sys.stderr, "    [%s%s]  #%s/%s%s\r" %(s1,s2,n,max,e),
+
+def clearprogress():
+    print >>sys.stderr, ' '*120+'\r',
+
+def log(a,verbose=0,debug=0,clear=True):
+    #print DEBUG,debug,VERBOSE,verbose
+    if type(a) is unicode:
+        a = a.encode(errors='ignore')
+    if debug:
+        if DEBUG >= debug:
+            if clear:
+                clearprogress()
+            print '>>>',a
+    elif VERBOSE >= verbose:
+        if clear:
+            clearprogress()
+        print a
+
+def error(a):
+    log('    *** ' + a)
 
 
 def lprint(e):
@@ -44,38 +76,40 @@ def cleantitle(tree):
     return text
 
 
-def progress(n,max,e=''):
-
-    l=80
-    p = n*80/max
-    s1 = '=' * p
-    s2 = ' ' * (l-p)
-    print "    [%s%s]  #%s/%s%s\r" %(s1,s2,n,max,e),
+def filename_pagecache(page):
+    return os.path.join(TMPDIR, 'page_%04i.html' %(page))
 
 
-def clear():
-    print ' '*100+'\r',
+def download_page(page,url=None,use_cache=True):
 
+    fname = filename_pagecache(page)
 
-def download_page(page,url=None):
-
-    if not url:
-        url = base + 'showthread.php?t=' + str(threadid)
-        if page:
-            url += '&page=%s' %(page)
-
-    if not USE_TMP:
-        data = urllib2.urlopen(url).read()
-    else:
-        with open(SLURP + 'page_%03i.html' %(page),'r') as f:
+    if use_cache and os.path.exists(fname):
+        log('Reading page %s from %s' %(page,fname),debug=1)
+        with open(fname,'r') as f:
             data = f.read()
 
+    else:
+        if not url:
+            url = base + 'showthread.php?t=' + str(threadid)
+            if page:
+                url += '&page=%s' %(page)
+
+        log('Download page %s from %s' %(page,url),debug=1)
+
+        try:
+            req = urllib2.urlopen(url)
+            data = req.read()
+
+            # Save to cache
+            with open(fname, 'wb') as f:
+                f.write(data)#.encode('utf-8'))
+
+        except urllib2.HTTPError as e:
+            error("%s: Failed to fetch %s" %(e.code,url))
+            raise
+
     tree = html.fromstring(data)
-
-    if not USE_TMP:
-        with open(SLURP + 'page_%03i.html' %(page), 'wb') as f:
-            f.write(data)
-
     return tree
 
 
@@ -97,35 +131,93 @@ class FakeUrl(object):
         return self.header
 
 
-def download_attachment(att):
-
-    url = base + 'attachment.php?attachmentid=' + str(att)
-
-    if not USE_TMP:
-        req = urllib2.urlopen(url)
-
-        with open(SLURP + str(att), 'wb') as f:
-            f.write(str(req.info()))
-
-        return req
-
-    else:
-
-        with open(SLURP + str(att), 'rb') as f:
-            header = f.read()
-            req = FakeUrl(header)
-
-        return req
+def filename_attcache(att):
+    return os.path.join(TMPDIR, str(att))
 
 
-def parse_page(tree):
+def get_attachment_cache(att):
+
+    fname = filename_attcache(att)
+
+    if not USE_CACHE:
+        return None
+    if not os.path.exists(fname):
+        return None
+
+    # Read header data from cache
+    with open(fname,'rb') as f:
+        header = {}
+        for l in f:
+            s=l.split(':')
+            if len(s)==2:
+                header[s[0].strip()] = s[1].strip()
+
+    # Get the actual filename
+    filename = str(att)
+    cd = header.get('Content-disposition','')
+    m=re.search(r'filename="(.*)"', cd)
+    if m:
+        filename=urllib.unquote(m.group(1)).decode('utf-8')
+
+    # Try ATT_filename first then filename
+    cache = os.path.join('attachments', str(att) + '_' + filename)
+    if not os.path.exists(cache):
+        cache = os.path.join('attachments', filename)
+        if not os.path.exists(cache):
+            return None
+
+    size = os.path.getsize(cache)
+    if size != int(header['Content-Length']):
+        return None
+
+    return cache
+
+
+def download_attachment(att,url):
+
+    # Open connection
+    req = urllib2.urlopen(url)
+    header = req.info()
+
+    # Store header data
+    with open(filename_attcache(att), 'wb') as f:
+        f.write(str(req.info()))
+
+    # Get the attachment info from the header
+    cd = header.getheader('Content-disposition')
+    m=re.search(r'filename="(.*)"', cd)
+    if m:
+        filename=urllib.unquote(m.group(1)).decode('utf-8')
+    filename = os.path.join('attachments', str(att) + '_' + filename)
+
+    log('Download attachment %s: %s' %(att,os.path.basename(filename)),debug=1)
+
+    # Read data from the web
+    data = req.read()
+
+    length = int(header.getheader('Content-Length'))
+    if len(data) != length:
+            raise Exception("Missing data from server/file, want %s bytes, got %s" %(length,len(data)))
+
+    # Write cache data
+    with open(filename,'wb') as f:
+        f.write(data)
+
+    return filename
+
+
+
+
+def parse_page(page,tree):
 
     posts = []
 
     #
     for e in tree.xpath('//ol[@id="posts"]/li'):
 
-        data = {}
+        data = {
+            'page': page,
+        }
 
         # POST ID
         t = e.get('id')
@@ -135,43 +227,35 @@ def parse_page(tree):
         m = re.search(r'post_(.*)', t)
         if not m:
             raise Exception("Missing post id")
-        post = int(m.group(1))
-        data['id'] = post
+        data['id'] = int(m.group(1))
 
         # POST COUNTER #1, #2, ...
         t = findclass(e,'.//a','postcounter')
         if not len(t):
             raise Exception("Missing post number")
-        counter = t[0].text
-        data['num'] = counter
+        data['num'] = t[0].text
 
         # TIME AND DATE
         t = findclass(e,'.//span','date')
         if not len(t):
             raise Exception("Missing date")
-        date = t[0].text
-        data['date'] = date
+        data['date'] = t[0].text
 
         t = findclass(e,'.//span','time')
         if not len(t):
             raise Exception("Missing time")
-        time = t[0].text
-        data['time'] = time
+        data['time'] = t[0].text
 
         # USERNAME
         t = findclass(e,'.//a','username')
         if not len(t):
             raise Exception("Missing username")
-        user = t[0].find('strong').text
-        data['user'] = user
+        data['user'] = t[0].find('strong').text
 
         # TITLE (OPTIONAL)
-        ptitle = None
         t = findclass(e,'.//h2','title')
         if len(t):
-            # Wash title
-            ptitle = cleantitle(t[0])
-            data['title'] = ptitle
+            data['title'] = cleantitle(t[0])
 
         # MAIN POST
         t = findclass(e, './/blockquote', 'postcontent')
@@ -179,56 +263,67 @@ def parse_page(tree):
             raise Exception("Missing post text")
         data['main'] = t[0]
 
-        #print "********************"
-        #print "NR   :",counter
-        #print "POST :",post
-        #print "DATE :",date
-        #print "TIME :",time
-        #print "USER :",user
-        #if ptitle:
-        #        print "TITLE:",ptitle
-
         posts.append(data)
 
     return posts
 
 
-print "vGrab v1.0 -- vBulletin thread grabber\nCopyright (C) 2015 Svein Seldal <sveinse@seldal.com>\nLicensed under GPL3.0\n"
+
+#-----------------------------------------------------------------------------
+ap = argparse.ArgumentParser()
+ap.add_argument('-t', '--tmpdir', help="Temp work directory", default='tmp')
+ap.add_argument('-n', '--nocache', help='Disable caching', action='store_true')
+ap.add_argument('-D', '--debug', help='Enable debugging', action='count', default=0)
+ap.add_argument('-v', '--verbose', help='Verbose output', action='count', default=0)
+ap.add_argument('url', help="URL to download", nargs='?', default=START_URL)
+
+opts = ap.parse_args()
+TMPDIR = opts.tmpdir
+DEBUG = opts.debug
+VERBOSE = opts.verbose
+USE_CACHE = not opts.nocache
+
+log('''vGrab v1.0 -- vBulletin thread grabber
+Copyright (C) 2015 Svein Seldal <sveinse@seldal.com>
+Licensed under GPL3.0
+''')
 
 
-print "\nDownloading first page..."
+
+#-----------------------------------------------------------------------------
+log("\nDownloading first page...")
 
 # Ensure tmp dir
-if not os.path.exists(SLURP):
-    os.makedirs(SLURP)
+if not os.path.exists(TMPDIR):
+    os.makedirs(TMPDIR)
 
-# Hent inn forste side
-tree = download_page(0,START_URL)
+# Load first page
+tree = download_page(0,opts.url,use_cache=False)
 
-# base
+# BASE URL for the site
 t=tree.xpath('//head/base')[0]
 base=t.attrib['href']
-print "    BASE:", base
+log("    BASE: " + base)
 
-# Tittel og ID
+# Title and thread ID
 t=tree.xpath('//*[@id="pagetitle"]/h1/span/a')[0]
 title=t.text
-print "    TITLE:", title
+log("    TITLE: " + title)
 
 m=re.search('/(\d+)-.*', t.attrib['href'])
 if m:
     threadid=int(m.group(1))
-print "    THREAD ID:", threadid
+log("    THREAD ID: " + str(threadid))
 
-# Hvor mange innlegg siden har
+# How many posts
 post_count = 0
 t=tree.xpath('//*[@id="postpagestats_above"]')[0].text
 m = re.search('av (\d+)', t)
 if m:
     post_count = int(m.group(1))
-print "    POSTS:", post_count
+log("    POSTS: " + str(post_count))
 
-# Hvor mange sider er posten paa
+# How many pages
 pages = 0
 for e in findclass(tree, '//a', 'popupctrl'):
     l = e.text
@@ -237,28 +332,38 @@ for e in findclass(tree, '//a', 'popupctrl'):
     m = re.search('Side \d+ av (\d+)', l)
     if m:
         pages = int(m.group(1))
-current_page = 0
-print "    PAGES:", pages
+log("    PAGES: " + str(pages))
+
+# How many do we have in our cache
+use_cache = [ USE_CACHE and os.path.exists(filename_pagecache(page))
+              for page in range(pages+1) ]
+
+# Always redownload the last page
+prev = False
+for n in range(pages,0,-1):
+    if prev == False and use_cache[n] == True:
+        use_cache[n] = False
+        break
+    prev = use_cache[n]
 
 
 
-print "\nDownloading posts..."
+#-----------------------------------------------------------------------------
+log("\nDownloading posts...")
 num = 0
 
 post_list = []
 posts = {}
 
-while current_page < pages:
+current_page = 1
+while current_page <= pages:
 
-    #print "===================="
-    #print "FETCHING PAGE:",current_page+1
+    tree = download_page(current_page,use_cache=use_cache[current_page])
 
-    tree = download_page(current_page+1)
-
-    for data in parse_page(tree):
+    for data in parse_page(current_page,tree):
 
         num += 1
-        progress(num, post_count, ', page %s/%s' %(current_page+1,pages))
+        progress(num, post_count, ', page %s/%s' %(current_page,pages))
 
         iid = data['id']
         if iid in posts:
@@ -269,11 +374,12 @@ while current_page < pages:
 
     current_page += 1
 
-print
+log('',clear=False)
 
 
 
-print "\nParsing links and images..."
+#-----------------------------------------------------------------------------
+log("\nParsing links and images...")
 num = 0
 
 attachment_fetchlist = {}
@@ -287,13 +393,12 @@ for post in post_list:
     data = posts[post]
     main = data['main']
 
-    # Find all images
+    # Find all images we want to download
     images = main.xpath('.//img')
     for img in images:
         src=img.get('src')
 
-        #clear()
-        #print '   ',src
+        log('IMG  ' + src, debug=2)
 
         m = re.search(r'attachmentid=(\d+)($|&)', src)
         if m:
@@ -307,67 +412,81 @@ for post in post_list:
     for link in links:
         href = link.get('href')
 
-        #clear()
-        #print '   ',href.encode('utf-8')
+        log('LINK ' + href, debug=2)
 
-        m=re.search('/(\d+)-.*post(\d+)$', href)
+        # Search for thread post links
+        m=re.search(r'/(\d+)-.*post(\d+)$', href)
         if m:
             tid = int(m.group(1))
             pid = int(m.group(2))
 
-            if tid != threadid:
-                raise Exception("Unknown threadid %s" %(tid))
-            if pid not in posts:
-                raise Exception("Unknown postid %s" %(pid))
+            # Proceed if link point to a post in our thread
+            if tid == threadid and pid in posts:
+                continue
 
-            continue
-
+        # Links that points to attachments
         m = re.search(r'attachmentid=(\d+)($|&)', href)
         if m:
             iid=int(m.group(1))
             attachment_fetchlist[iid]=post
-
             continue
 
-        clear()
-        print '   ',href.encode('utf-8')
+        log('    ' + href,verbose=1)
 
-print
 
 # Manual adds
-image_fetchlist['images/misc/quote-left.png'] = True
+for img in EXTRA_IMAGES:
+    if img not in image_fetchlist:
+        image_fetchlist[img] = True
 
-print "\nDownloading icons..."
+log('',clear=False)
+
+
+
+#-----------------------------------------------------------------------------
+log("\nDownloading images...")
 num = 0
 maxnum = len(image_fetchlist)
+failed = 0
 
 for image in image_fetchlist:
 
     num += 1
     progress(num, maxnum)
 
-    path = os.path.split(image)
-    if not os.path.exists(path[0]):
-        os.makedirs(path[0])
+    if USE_CACHE and os.path.exists(image):
+        log('Skipping image %s' %(image),debug=1)
 
-    url = base + image
-    #print url
-    try:
-        req = urllib2.urlopen(url)
-        #print req.info()
-        with open(image,'wb') as f:
-            f.write(req.read())
+    else:
+        url = base + image
+        log('Download image %s' %(url),debug=1)
 
-    except urllib2.HTTPError as e:
-        clear()
-        print "    %s: Failed to fetch" %(e.code),url
+        try:
+            req = urllib2.urlopen(url)
+            data = req.read()
+
+            # Ensure dir exists
+            path = os.path.split(image)
+            if not os.path.exists(path[0]):
+                os.makedirs(path[0])
+
+            # Save image
+            with open(image,'wb') as f:
+                f.write(data)
+
+        except urllib2.HTTPError as e:
+            error("%s: Failed to fetch %s" %(e.code,url))
+            failed += 1
+
+log('',clear=False)
+
+if failed:
+    log('    FAILED TO DOWNLOAD: ' + str(failed))
 
 
-print
 
-
-
-print "\nDownloading attachments..."
+#-----------------------------------------------------------------------------
+log("\nDownloading attachments...")
 
 if not os.path.exists('attachments'):
     os.makedirs('attachments')
@@ -384,72 +503,51 @@ for (att,post) in attachment_fetchlist.items():
     num += 1
     progress(num, maxnum)
 
-    pnum = posts[post]['num']
+    log('ATT ' + str(att), debug=2)
 
-    try:
-        req = download_attachment(att)
+    # Is the cache valid? -- The cache will never be valid if
+    # USE_CACHE is false
+    filename = get_attachment_cache(att)
 
-        header = req.info()
-        length = int(header.getheader('Content-Length'))
-        #ctype = header.getheader('Content-Type')
+    if filename:
+        # Cache is OK
+        log('Skipping attachment %s: %s' %(att,os.path.basename(filename)),debug=1)
 
-        disp = header.getheader('Content-disposition')
-        m=re.search(r'filename="(.*)"', disp)
-        if not m:
-            raise Exception("Missing filename")
-        filename = urllib.unquote(m.group(1)).decode('utf-8')
+    else:
+        try:
 
-        if filename in filenames:
-            filename = str(att) + '_' + filename
-        if att in attachments:
-            raise Exception("Attachment %s already fetched" %(att))
+            # Download the attachment
+            url = base + 'attachment.php?attachmentid=' + str(att)
+            filename = download_attachment(att,url)
+
+        except urllib2.HTTPError as e:
+            pnum = posts[post]['num']
+            error("%s: Failed to fetch attachment %s for post %s" %(e.code,att,pnum))
+            failed += 1
+
+    if filename:
+
+        # Store filenames and attachment names
         filenames[filename] = att
         attachments[att] = filename
 
-        fname = 'attachments/' + filename
-
-        if not os.path.exists(fname):
-            if not USE_TMP:
-                data = req.read()
-            else:
-                # Hack. Should rather retry downloading the file
-                clear()
-                print "    *** Missing local file "+fname
-                data = ' '*length
-
-        else:
-            with open(fname,'rb') as f:
-                data=f.read()
-
-        if len(data) != length:
-            raise Exception("Missing data from server/file, want %s bytes, got %s" %(length,len(data)))
-
-        if not USE_TMP:
-            with open(fname, 'wb') as f:
-                f.write(data)
-
-        #else:
-        #    clear()
-        #    print "    %s: Skipping" %(fname)
-
-    except urllib2.HTTPError as e:
-        clear()
-        print "    *** Failed to fetch (%s) attachment %s in post %s" %(e.code,att,pnum)
-        failed += 1
-
-    except IOError as e:
-        clear()
-        print "    *** Missing local attachment %s in post %s" %(att,pnum)
-        failed += 1
-
 progress(num, maxnum)
-print
+log('',clear=False)
 
 if failed:
-    print '    FAILED TO DOWNLOAD:',failed
+    log('    FAILED TO DOWNLOAD: ' +str(failed))
 
 
 
+
+
+
+
+
+sys.exit(1)
+
+
+#-----------------------------------------------------------------------------
 print "\nCreating web data..."
 num = 0
